[CI/Build] Reorganize models tests
