+++ tests/models/language/generation/test_granite.py
