+++ tests/models/multimodal/generation/test_awq.py
