+++ tests/models/multimodal/generation/test_broadcast.py
