+++ tests/models/multimodal/generation/test_common.py
