+++ tests/models/multimodal/generation/test_florence2.py
