+++ tests/models/multimodal/generation/test_granite_speech.py
