+++ tests/models/multimodal/generation/test_intern_vit.py
