+++ tests/models/multimodal/generation/test_mllama.py
