+++ tests/models/multimodal/generation/test_phi4mm.py
