+++ tests/models/multimodal/generation/test_pixtral.py
