+++ tests/models/multimodal/generation/test_qwen2_vl.py
