+++ tests/models/multimodal/generation/test_ultravox.py
