+++ tests/models/multimodal/generation/test_whisper.py
