+++ tests/models/multimodal/generation/vlm_utils/builders.py
