+++ tests/models/multimodal/generation/vlm_utils/case_filtering.py
