+++ tests/models/multimodal/generation/vlm_utils/core.py
