+++ tests/models/multimodal/generation/vlm_utils/custom_inputs.py
