+++ tests/models/multimodal/generation/vlm_utils/model_utils.py
