+++ tests/models/multimodal/generation/vlm_utils/runners.py
