+++ tests/models/multimodal/generation/vlm_utils/types.py
